requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
Flask==3.1.2
Flask-SocketIO==5.5.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pygame==2.6.1
python-engineio==4.12.2
python-socketio==5.13.0
requests==2.32.5
simple-websocket==1.1.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
